refactor: validator 함수 분리
