+++ hooks/src/lib/cvc/useCVC.ts
@@ -1,6 +1,7 @@
 import { ChangeEvent, useState } from 'react';
 import { ValidationResult } from '../types';
 import { CVC_ERROR_TYPES, ERROR_MESSAGE } from '../constants';
+import { checkIsNumber, checkIsValidLength } from '../validators';
 
 function useCVC() {
   const [CVC, setCVC] = useState('');
@@ -10,18 +11,9 @@ function useCVC() {
     errorMessage: '',
   });
 
-  const checkIsNumber = (value: string) => {
-    const regex = /^[0-9]*$/;
-    return regex.test(value);
-  };
-
-  const checkIsValidLength = (value: string) => {
-    return value.length <= 3;
-  };
-
   const validateCVC = (value: string) => {
     const isNumber = checkIsNumber(value);
-    const isValidLength = checkIsValidLength(value);
+    const isValidLength = checkIsValidLength(value, 3);
 
     if (!isNumber) {
       return CVC_ERROR_TYPES.notNumber;
